// free-todo-frontend/lib/generated/schemas/activityListResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LifeTrace API
 * 智能生活记录系统 API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityResponse } from './activityResponse';

export interface ActivityListResponse {
  activities: ActivityResponse[];
  total_count: number;
}