// free-todo-frontend/lib/generated/schemas/eventDetailResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * LifeTrace API
 * 智能生活记录系统 API
 * OpenAPI spec version: 0.1.0
 */
import type { EventDetailResponseAppName } from './eventDetailResponseAppName';
import type { EventDetailResponseWindowTitle } from './eventDetailResponseWindowTitle';
import type { EventDetailResponseEndTime } from './eventDetailResponseEndTime';
import type { ScreenshotResponse } from './screenshotResponse';
import type { EventDetailResponseAiTitle } from './eventDetailResponseAiTitle';
import type { EventDetailResponseAiSummary } from './eventDetailResponseAiSummary';

export interface EventDetailResponse {
  id: number;
  app_name: EventDetailResponseAppName;
  window_title: EventDetailResponseWindowTitle;
  start_time: string;
  end_time: EventDetailResponseEndTime;
  screenshots: ScreenshotResponse[];
  ai_title?: EventDetailResponseAiTitle;
  ai_summary?: EventDetailResponseAiSummary;
}